Moved demo to separate module, replaced cursors for mag and move.
